+++ services/madoc-ts/src/extensions/capture-models/ProjectSelector/ProjectSelector.preview.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { LocaleString } from '../../../frontend/shared/components/LocaleString';
+import { Heading5 } from '../../../frontend/shared/typography/Heading5';
+import { ProjectSelectorProps } from './ProjectSelector';
+
+export const ProjectSelectorPreview: React.FC<ProjectSelectorProps> = props => {
+  if (props.value) {
+    return <Heading5 as={LocaleString}>{props.value.label}</Heading5>;
+  }
+
+  return null;
+};