+++ services/madoc-ts/src/extensions/capture-models/ProjectSelector/ProjectSelector.tsx
@@ -1,39 +1,38 @@
+import { InternationalString } from '@hyperion-framework/types';
 import React, { useRef } from 'react';
 import { useInfiniteQuery } from 'react-query';
+import { LocaleString } from '../../../frontend/shared/components/LocaleString';
 import { Button } from '../../../frontend/shared/navigation/Button';
-import { Heading5 } from '../../../frontend/shared/typography/Heading5';
-import { ImageGrid } from '../../../frontend/shared/atoms/ImageGrid';
-import { CroppedImage } from '../../../frontend/shared/atoms/Images';
-import { ImageStripBox } from '../../../frontend/shared/atoms/ImageStrip';
+import { Heading3 } from '../../../frontend/shared/typography/Heading3';
 import { useApi } from '../../../frontend/shared/hooks/use-api';
 import { useInfiniteAction } from '../../../frontend/site/hooks/use-infinite-action';
 
-export type MediaExplorerProps = {
+export type ProjectSelectorProps = {
   id: string;
   label: string;
   type: string;
   value: {
     id: string;
-    image: string;
-    thumbnail: string;
+    label: InternationalString;
   } | null;
 };
 
-export const MediaExplorer: React.FC<MediaExplorerProps & {
-  updateValue: (value: MediaExplorerProps['value']) => void;
+export const ProjectSelector: React.FC<ProjectSelectorProps & {
+  updateValue: (value: ProjectSelectorProps['value']) => void;
 }> = props => {
   const api = useApi();
   const container = useRef<HTMLDivElement>(null);
   const { data: pages, fetchMore, canFetchMore, isFetchingMore } = useInfiniteQuery(
-    ['media-explorer', {}],
+    ['get-projects-list', {}],
     async (key, _, vars: { page?: number }) => {
-      return api.media.listMedia(vars.page);
+      return api.getProjects(vars?.page);
     },
     {
       getFetchMore: lastPage => {
         if (lastPage.pagination.totalPages === lastPage.pagination.page) {
           return undefined;
         }
+
         return {
           page: lastPage.pagination.page + 1,
         };
@@ -48,33 +47,24 @@ export const MediaExplorer: React.FC<MediaExplorerProps & {
   });
 
   if (props.value) {
-    const chosenMedia = props.value;
     return (
       <div>
-        <ImageStripBox>
-          <CroppedImage>{chosenMedia.thumbnail ? <img src={chosenMedia.thumbnail} alt="thumb" /> : null}</CroppedImage>
-        </ImageStripBox>
-        <Button onClick={() => props.updateValue(null)}>Choose another image</Button>
+        <Heading3 as={LocaleString}>{props.value.label}</Heading3>
+        <Button onClick={() => props.updateValue(null)}>Choose another project</Button>
       </div>
     );
   }
 
   return (
     <div ref={container} style={{ maxHeight: 500, overflowY: 'scroll' }}>
-      <ImageGrid>
+      <div>
         {pages?.map((page, key) => {
           return (
             <React.Fragment key={key}>
-              {page.mediaItems.map(media => (
-                <ImageStripBox
-                  key={media.id}
-                  onClick={() =>
-                    props.updateValue({ id: media.id, image: media.publicLink, thumbnail: media.thumbnail })
-                  }
-                >
-                  <CroppedImage>{media.thumbnail ? <img src={media.thumbnail} alt="thumb" /> : null}</CroppedImage>
-                  <Heading5>{media.displayName}</Heading5>
-                </ImageStripBox>
+              {page.projects.map(project => (
+                <div key={project.id} onClick={() => props.updateValue({ id: project.slug, label: project.label })}>
+                  <Heading3 as={LocaleString}>{project.label}</Heading3>
+                </div>
               ))}
             </React.Fragment>
           );
@@ -82,7 +72,7 @@ export const MediaExplorer: React.FC<MediaExplorerProps & {
         <Button ref={loadMoreButton} onClick={() => fetchMore()} style={{ display: canFetchMore ? 'block' : 'none' }}>
           Load more
         </Button>
-      </ImageGrid>
+      </div>
     </div>
   );
 };