+++ services/madoc-ts/src/extensions/page-blocks/default-definitions.ts
@@ -1,4 +1,5 @@
 import { CanvasPanelBlock } from '../../frontend/shared/components/CanvasPanelBlock';
+import { SingleProject } from '../../frontend/shared/components/SingleProject';
 import { Surface } from '../../frontend/shared/layout/Surface';
 import { CrowdSourcingBanner } from '../../frontend/shared/components/CrowdSourcingBanner';
 import { PageBlockDefinition } from './extension';
@@ -12,5 +13,6 @@ export const defaultPageBlockDefinitions: PageBlockDefinition<any, any, any, any
   (Surface as any)[Symbol.for('slot-model')] as any,
   (CrowdSourcingBanner as any)[Symbol.for('slot-model')] as any,
   (CanvasPanelBlock as any)[Symbol.for('slot-model')] as any,
+  (SingleProject as any)[Symbol.for('slot-model')] as any,
   simpleMarkdownBlock,
 ];