+++ services/madoc-ts/src/frontend/shared/required-modules.ts
@@ -1,0 +1,1 @@
+import './capture-models/editor/bundle';