+++ services/madoc-ts/src/types/get-slots.ts
@@ -1,0 +1,6 @@
+import { BlockCollector } from './block-collector';
+import { EditorialContext } from './schemas/site-page';
+
+export type GetSlots = (ctx: EditorialContext, options?: GetSlotsOptions) => Promise<any> | any;
+
+export type GetSlotsOptions = { slotIds?: string[]; collector?: BlockCollector };