+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "tiny" variation of image wrapper [dev]
 - Added preview for entity lists using selector if it's available
 - Added "confirm" to deselect or confirm the current selector
+- Added support for API calls as part of a page block (with server rendering) [dev]
+- Added "Single project" page block
 
 ### Changed
 - The link in the top bar now always links to the site (previously the admin for admins).